Rename namespace to DependencyInjection
