+++ Cql/CoreTests/ExpressionBuilderTests.cs
@@ -4,7 +4,7 @@ using System;
 using System.IO;
 using System.Linq;
 using Hl7.Cql.Compiler;
-using Hl7.Cql.Compiler.Services;
+using Hl7.Cql.Compiler.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace CoreTests
