+++ Cql/CoreTests/QueriesTest.cs
@@ -9,7 +9,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using Hl7.Cql.Compiler;
-using Hl7.Cql.Compiler.Services;
+using Hl7.Cql.Compiler.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace CoreTests
