+++ Cql/CqlToElmTests/XmlTest.cs
@@ -9,7 +9,7 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Xml.Serialization;
-using Hl7.Cql.CqlToElm.Services;
+using Hl7.Cql.CqlToElm.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Hl7.Cql.CqlToElm.Test
